Optionally support apache config for experimenting with prod routing rules
Summary:
In production, we want googlebot to be able to crawl the site and we want
individual post links to work. This patch has a hacked together apache config so
we can test locally before making real updates to the rdio chef repo.

Test Plan: Ran apache as it was installed on my machine. It worked.

Reviewers: endenizen, jesse, bryan.hughes

Reviewed By: bryan.hughes

CC: amie.kweon

Differential Revision: https://phabricator.rdio.com/D11119